admin/requirements/base.txt
psycopg2==2.9.5
cron-validator==1.0.0
Jinja2==3.1.2
nh3==0.2.14

# DJANGO
django==3.2
django-split-settings==1.2.0
django-ckeditor==6.7.0
django-cronfield==0.2
